requirements/base.txt

# Django
# ------------------------------------------------------------------------------
Django==3.0.7  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5 # https://github.com/joke2k/django-environ
django-extensions==2.1.6  # https://github.com/django-extensions/django-extensions

# Manage the CORS permissions
django-cors-headers==3.4.0	# https://pypi.org/project/django-cors-headers/

# Django REST Framework
djangorestframework==3.10.3  # https://github.com/encode/django-rest-framework
djangorestframework-jwt==1.11.0		# http://getblimp.github.io/django-rest-framework-jwt/
coreapi==2.3.3	# https://github.com/core-api/python-client
django-rest-framework==0.1.0
coreschema==0.0.4
drf-yasg==1.17.1

# postgrest
# ------------------------------------------------------------------------------
psycopg2-binary==2.8.5  # https://github.com/psycopg/psycopg2

# Code quality
# ------------------------------------------------------------------------------
pylint-django==2.0.9  # https://github.com/PyCQA/pylint-django

# Custom dependencies
# ------------------------------------------------------------------------------
idna==2.10
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
simplejson==3.16.0
sqlparse==0.3.1
uritemplate==3.0.0
urllib3==1.25.10
autopep8==1.4.4
certifi==2020.6.20
chardet==3.0.4
asgiref==3.2.10
